+++ src/main/kotlin/lotto/domain/LottoShop.kt
@@ -1,5 +1,7 @@
 package lotto.domain
 
+import lotto.vo.Money
+
 class LottoShop(private val lottoMachine: LottoMachine = DefaultLottoMachine) {
     fun buying(amount: Money): LottoTickets {
         val ticketCount = amount / LOTTO_TICKET_PRICE
