+++ src/test/kotlin/lotto/vo/MoneyTest.kt
@@ -1,10 +1,11 @@
-package lotto.domain
+package lotto.vo
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.inspectors.forAll
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
+import lotto.vo.Money
 import java.math.BigDecimal
 import java.math.RoundingMode
